# backend/openapi.yaml
openapi: 3.0.0
info:
  title: 'Chiorino ERP API'
  version: '1'
paths:
...
  /bp-groups:
    get:
      tags:
        - 'BP Groups'
      summary: 'Get list of bp groups'
      operationId: e49f2eb6309b47f382b9cbc1c2a2d477
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/BPGroupResource' } }
                type: object
    post:
      tags:
        - 'BP Groups'
      summary: 'Create new bp group'
      operationId: b62e02ab8fd4a2075ea7687b2d4d184b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
              type: object
            example:
              name: Exclusive
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/BPGroupResource' } }
                type: object
  '/bp-groups/{id}':
    get:
      tags:
        - 'BP Groups'
      summary: 'Get details of bp group'
      operationId: 025e5e02146a5abd5f8f0981154e6f06
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/BPGroupResource' } }
                type: object
    put:
      tags:
        - 'BP Groups'
      summary: 'Update bp group'
      operationId: aad2e3261809c7f29b76deeb71327807
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
              type: object
            example:
              name: Exclusive
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/BPGroupResource' } }
                type: object
    delete:
      tags:
        - 'BP Groups'
      summary: 'Delete bp group'
      operationId: 8d32c260dc24cbee90473520b4a1d868
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/BPGroupResource' } }
                type: object
...
  /contacts:
    get:
      tags:
        - Contacts
      summary: 'Get list of contacts'
      operationId: 7b8056a6c305d3995f6d8cc46d3cae65
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ContactResource' } }
                type: object
    post:
      tags:
        - Contacts
      summary: 'Create new contact'
      operationId: 292652dc2e92a1f2c52aa7c864b8f718
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - type
                - name
                - language
                - is_employee
              properties:
                type:
                  type: string
                name:
                  type: string
                note:
                  type: string
                contact_type_id:
                  type: integer
                department:
                  type: string
                address_id:
                  type: string
                office_phone:
                  type: string
                mobile_phone:
                  type: string
                email:
                  type: string
                language:
                  type: string
                is_employee:
                  type: boolean
              type: object
            example:
              type: person
              name: John
              note: note
              contact_type_id: 1
              department: Production
              address_id: null
              office_phone: null
              mobile_phone: null
              email: null
              language: en
              is_employee: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
  '/contacts/{id}':
    get:
      tags:
        - Contacts
      summary: 'Get details of single contact'
      operationId: 371008c3a140dc687487e65c833b4c05
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
    put:
      tags:
        - Contacts
      summary: 'Update contact'
      operationId: 988ee956022c57234f5b5871a911543d
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - type
                - name
                - language
                - is_employee
              properties:
                type:
                  type: string
                name:
                  type: string
                note:
                  type: string
                contact_type_id:
                  type: integer
                department:
                  type: string
                address_id:
                  type: string
                office_phone:
                  type: string
                mobile_phone:
                  type: string
                email:
                  type: string
                language:
                  type: string
                is_employee:
                  type: boolean
              type: object
            example:
              type: person
              name: John
              note: note
              contact_type_id: 1
              department: Production
              address_id: null
              office_phone: null
              mobile_phone: null
              email: null
              language: en
              is_employee: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
    delete:
      tags:
        - Contacts
      summary: 'Delete contact'
      operationId: aa6804ea70c88459eaa3b88c8bab0195
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContactResource'
        '403':
          description: Unauthorized
        '500':
          description: 'Internal Error'
...